Update Docker configuration for better file management
Updated Dockerfile copy commands syntax and added .dockerignore files for both frontend and backend in order to ignore unnecessary files such as node_modules, temporary files and configuration files. This will help to stop these files from being copied into the Docker image which reduces the size of the image and decreases build time. Also, added volumes in docker-compose.yml to share the code between the host and the Docker containers for the backend and frontend for improved development workflow. The containers now depend on the db service too.